refactor: move core partition mappers to a new package
