+++ airflow-core/src/airflow/timetables/simple.py
@@ -16,7 +16,6 @@
 # under the License.
 from __future__ import annotations
 
-from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING, Any, TypeAlias
 
 from airflow._shared.timezones import timezone
@@ -34,10 +33,11 @@ except ModuleNotFoundError:
 
 
 if TYPE_CHECKING:
-    from collections.abc import Collection, Iterable, Sequence
+    from collections.abc import Collection, Sequence
 
     from pendulum import DateTime
 
+    from airflow.partition_mapper.base import PartitionMapper
     from airflow.timetables.base import TimeRestriction
     from airflow.utils.types import DagRunType
 
@@ -263,36 +263,3 @@ class PartitionedAssetTimetable(AssetTriggeredTimetable):
             assets=decode_asset_like(data["asset_condition"]),
             partition_mapper=decode_partition_mapper(data["partition_mapper"]),
         )
-
-
-class PartitionMapper(ABC):
-    """
-    Base partition mapper class.
-
-    Maps keys from asset events to target dag run partitions.
-    """
-
-    @abstractmethod
-    def to_downstream(self, key: str) -> str:
-        """Return the target key that the given source partition key maps to."""
-
-    @abstractmethod
-    def to_upstream(self, key: str) -> Iterable[str]:
-        """Yield the source keys that map to the given target partition key."""
-
-    def serialize(self) -> dict[str, Any]:
-        return {}
-
-    @classmethod
-    def deserialize(cls, data: dict[str, Any]) -> PartitionMapper:
-        return cls()
-
-
-class IdentityMapper(PartitionMapper):
-    """Partition mapper that does not change the key."""
-
-    def to_downstream(self, key: str) -> str:
-        return key
-
-    def to_upstream(self, key: str) -> Iterable[str]:
-        yield key
